+++ History.md
@@ -1,7 +1,11 @@
 ## vNEXT
 
-## v1.1.1
+## v2.0.0
 * Add newer Meteor versions to `versionsFrom`
+* Minimum Meteor version is 2.8.1
+* onStartup add defer and default query [PR](https://github.com/Meteor-Community-Packages/meteor-user-status/pull/163) [@Sergeant61](https://github.com/Sergeant61)
+* Updated for Meteor 3
+* Updated to eslint v9
 
 ## v1.1.0